requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-telegram-bot==20.1
SQLAlchemy==2.0.37
pymysql==1.1.1
python-dotenv==1.0.0
pycryptodome==3.20.0
PyJWT==2.8.0
pydantic==2.8.2
requests==2.31.0
python-multipart==0.0.6
